+++ azd/src/main/java/org/azd/enums/AttachmentUploadType.java
@@ -1,0 +1,9 @@
+package org.azd.enums;
+
+/**
+ * Attachment upload type
+ */
+public enum AttachmentUploadType {
+    SIMPLE,
+    CHUNKED
+}